checkpoint/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
Pillow==8.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.4.1
toml==0.10.1
